Add build time version information.
This adds a TU containing information about the build and the current version.

This allows looking up the major/minor/revision, as well as the SHA of the
git commit where the build was run.

Inspired by the book "Professional CMake: A Practical Guide" by Craig Scott.
See https://crascit.com/professional-cmake/